Add files via upload
Designed a simple game comprising of minesweeper and tic tac toe. With modes of difficulty in minesweeper game a user can play at any difficulty. At first the user would have to register himself and can then play the games. There is a betting criteria in the game where the user can bet on a game. Moreover the system holds the progress of the user.